Move BinaryFormatterDataBusSerializer to obsolete file
